+++ CHANGELOG.md
@@ -1,3 +1,13 @@
-## 0.0.1
+# Changelog
 
-* TODO: Describe initial release.
+All notable changes to this project will be documented in this file.
+
+## 0.0.1
+- Initial release of `online_indicator` package.
+- Added `OnlineIndicator` widget to easily show an online/offline status badge.
+- Supports customization for:
+    - Background color
+    - Border color and width
+    - Radius (size) of the indicator
+    - Optional icon inside the indicator.
+- Lightweight and easy to integrate with profile images or any widgets.